agent/requirements.txt
langgraph
langgraph-prebuilt
langgraph-sdk
langgraph-checkpoint-sqlite
langsmith
langchain-community
langchain-core
langchain-openai
notebook
langchain-tavily
wikipedia
trustcall
langgraph-cli[inmem]
python-dotenv
upstash-redis
pypdf
faiss-cpu
langchain-text-splitters
redis
fastapi==0.104.1
starlette==0.27.0
uvicorn==0.23.2
psycopg2-binary
google-api-python-client
google-generativeai
pillow
python-multipart
